+++ src/entity/defines.rs
@@ -1,5 +1,5 @@
-pub const TILE_WIDTH: i32 = 4;
-pub const TILE_HEIGHT: i32 = 4;
+pub const TILE_WIDTH: u8 = 4;
+pub const TILE_HEIGHT: u8 = 4;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum TileState {
